PERF: Store hash of RsMacro and RsMacroCall in stubs
